[fix][shell] "write" command에서 Data에 0x7FFFFFFF 보다 큰 값을 문자열로 입력할 경우 exception 발생"
std::stoi() 함수가 signed int를 반환하는 함수이기 때문에
"0x7FFFFFFF"보다 큰 문자열을 입력할 경우 "out_of_range" exception을 발생하여
std::stoll() 함수로 변경하였습니다.

+++ TestShellApplication/WriteCommand.cpp
@@ -61,5 +61,5 @@ int WriteCommand::_checkAndGetData(string paramString)
 		throw std::exception("[Invalid Data] Not Hex Character");
 	}
 
-	return std::stoi(paramString, nullptr, 16);
+	return std::stoll(paramString, nullptr, 16);
 }
